import a PMOD pinout declaration library
